+++ json/cJSON.h
@@ -46,21 +46,17 @@ extern "C"
 /* The cJSON structure: */
 typedef struct cJSON {
 	struct cJSON *next,*prev;	/* next/prev allow you to walk array/object chains. Alternatively, use GetArraySize/GetArrayItem/GetObjectItem */
-	struct cJSON *child;		/* An array or object item will have a child pointer pointing to a chain of the items in the array/object. */
-
-	int type;					/* The type of the item, as above. */
-
-	char *valuestring;			/* The item's string, if type==cJSON_String */
+	struct cJSON *child;		  /* An array or object item will have a child pointer pointing to a chain of the items in the array/object. */
+	int type;					        /* The type of the item, as above. */
+	char *valuestring;			  /* The item's string, if type==cJSON_String */
   
-  // taka add
-  uint8_t valueuint8;
-  uint16_t valueuint16;
-  
-  
-	int valueint;				/* The item's number, if type==cJSON_Number */
-	double valuedouble;			/* The item's number, if type==cJSON_Number */
+  // taka add begin
+  uint8_t valueuint8;       /* The item's number, if type==cJSON_Number */
+  uint16_t valueuint16;     /* The item's number, if type==cJSON_Number */
 
-	char *string;				/* The item's name string, if this item is the child of, or is in the list of subitems of an object. */
+	int valueint;				      /* The item's number, if type==cJSON_Number */
+	double valuedouble;			  /* The item's number, if type==cJSON_Number */
+	char *string;				      /* The item's name string, if this item is the child of, or is in the list of subitems of an object. */
 } cJSON;
 
 typedef struct cJSON_Hooks {