+++ HISTORY.md
@@ -4,7 +4,7 @@
 - [x] rewrite test server in c, test cases in golang
 - [x] update travis ci with new test env
 - [x] implement timeout.set, timeout.get functions
-- [x] implement json_get_double, json_set_double, json_get_long, json_get_ulong
+- [x] implement json_get_double, json_set_double, json_get_long
 - [x] refactor tid data type
 - [x] refactor command definition